Move `KeyCounter` components to `HUD` namespace
