+++ osu.Game/Screens/Play/HUD/DefaultKeyCounterDisplay.cs
@@ -6,7 +6,7 @@ using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osuTK.Graphics;
 
-namespace osu.Game.Screens.Play
+namespace osu.Game.Screens.Play.HUD
 {
     public partial class DefaultKeyCounterDisplay : KeyCounterDisplay
     {
