+++ osu.Game/Screens/Play/HUD/KeyCounter.cs
@@ -5,7 +5,7 @@ using osu.Framework.Bindables;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 
-namespace osu.Game.Screens.Play
+namespace osu.Game.Screens.Play.HUD
 {
     public abstract partial class KeyCounter : Container
     {
